requirements.txt
absl-py==2.1.0
Adafruit-Blinka==8.39.1
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.0.0
adafruit-circuitpython-motor==3.4.13
adafruit-circuitpython-pca9685==3.4.15
adafruit-circuitpython-register==1.9.18
adafruit-circuitpython-requests==3.2.9
adafruit-circuitpython-servokit==1.3.17
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.63.0
Adafruit-PureIO==1.1.11
albucore==0.0.9
albumentations==1.4.8
annotated-types==0.7.0
appdirs==1.4.4
approxeng.input==2.6.4
apturl==0.5.2
asttokens==2.4.1
attrs==21.2.0
bcrypt==3.2.0
beniget==0.4.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cmake==3.29.3
colorama==0.4.4
coloredlogs==15.0.1
cpuset==1.6
cryptography==3.4.8
cupshelpers==1.0
cvzone==1.6.1
cycler==0.11.0
Cython==0.29.37
dbus-python==1.2.18
decorator==4.4.2
deep-sort-realtime==1.3.2
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
evdev==1.6.1
exceptiongroup==1.2.1
executing==2.0.1
fasteners==0.14.1
filelock==3.6.0
flatbuffers==24.3.25
fonttools==4.29.1
fs==2.4.12
fsspec==2024.3.1
future==0.18.2
gast==0.5.2
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.64.1
httplib2==0.20.2
humanfriendly==10.0
idna==3.3
imageio==2.34.1
importlib-metadata==4.6.4
ipython==8.25.0
jedi==0.19.1
jeepney==0.7.1
jetson-examples==0.1.4
jetson-stats==4.2.8
Jetson.GPIO==2.1.7
Jinja2==3.1.4
joblib==1.4.2
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.1
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
onnx==1.16.0
onnxruntime-gpu @ file:///home/dima/onnxruntime_gpu-1.17.0-cp310-cp310-linux_aarch64.whl
onnxsim==0.4.36
opencv-python==4.10.0.84
packaging==21.3
pandas==1.3.5
paramiko==2.9.3
parso==0.8.4
pexpect==4.8.0
pillow==10.3.0
platformdirs==2.5.1
ply==3.11
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycairo==1.20.1
pycocotools==2.0.7
pycuda==2024.1
pycups==2.0.1
pydantic==2.7.4
pydantic_core==2.18.4
pyftdi==0.55.4
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyparsing==2.4.7
pyRFC3339==1.1
pyserial==3.5
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-dbusmock==0.27.5
python-debian==0.1.43+ubuntu1.1
pythran==0.10.0
pytools==2024.1.5
pytz==2022.1
pyusb==1.2.1
pyxdg==0.27
PyYAML==6.0.1
requests==2.25.1
rich==13.7.1
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
SecretStorage==3.3.1
six==1.16.0
smbus2==0.4.3
smmap==5.0.1
stack-data==0.6.3
sympy==1.9
systemd-python==234
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorrt==8.6.2
tensorrt-dispatch==8.6.2
tensorrt-lean==8.6.2
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.5.22
tomli==2.0.1
torch @ file:///home/dima/torch-2.1.0-cp310-cp310-linux_aarch64.whl
torchvision==0.16.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.11.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
ultralytics==8.2.15
unicodedata2==14.0.0
urllib3==1.26.5
urwid==2.1.2
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.13
Werkzeug==3.0.3
xdg==5
xkit==0.0.0
zipp==1.0.0
